Merge CAS upload code from pkg/cas into pkg/client
This is a simple copy/paste/fix exercise to make the newer upload logic
accessible from the client package.

This adds more to the public interface of the Client struct as well as a
bunch of internal implementation logic.
The moved code is fairly isolated into its own files, except for struct
fields which are grouped together to simplify further refactoring.

Since this commit includes syntatic fixes (mostly renames) to satisfiy
the compiler, it's not completely risk free. However, any issues, if
any, would be present in the newly moved code.

Tested by building the repo and running existing tests.